move page to app + rename + fix tests
